pyproject.toml
[project]
name = "us-stock-research"
version = "0.1.0"
description = "M1 local bootstrap for US stock swing-trading research data."
requires-python = ">=3.11"
dependencies = [
  "google-cloud-storage==2.18.2",
  "httpx==0.27.2",
  "loguru==0.7.2",
  "notion-client==2.2.1",
  "pandas==2.2.3",
  "psycopg[binary]==3.2.3",
  "pyarrow==18.1.0",
  "pydantic-settings==2.6.1",
  "PyYAML==6.0.2",
  "SQLAlchemy==2.0.36",
  "tenacity==9.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff==0.8.3",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
